Namespace corrected and code refactored in LDAPUserProvided
